feat: add Ownership interface and update request validator

+++ backend/Yggdrasil.Application/Services/QuizService.cs
@@ -10,7 +10,8 @@ namespace Yggdrasil.Application.Services;
 public class QuizService(
     IQuizRepository quizRepository,
     ICategoryRepository categoryRepository,
-        ILogger<QuizService> logger
+        ILogger<QuizService> logger,
+        ICurrentUser currentUser
     ) : IQuizService
 {
     public async Task<QuizResponse> CreateQuizAsync(CreateQuizRequest request, CancellationToken cancellationToken)
@@ -27,7 +28,7 @@ public class QuizService(
             Id = Guid.NewGuid(),
             Title = request.Title,
             Description = request.Description,
-            OwnerId = Guid.NewGuid(), // will add ownsership interface
+            OwnerId =  currentUser.UserId,
             Difficulty = request.Difficulty,
             CreatedAt = DateTimeOffset.UtcNow,
             UpdatedAt = DateTimeOffset.UtcNow,
@@ -94,6 +95,11 @@ public class QuizService(
             logger.LogWarning("Quiz with id {id} not found", id);
             throw new NotFoundException("Quiz", id);
         }
+        if (quiz.OwnerId != currentUser.UserId)
+        {
+            throw new ForbiddenException("update this quiz");
+        }
+        
         var categories = await categoryRepository.GetByIdsAsync(request.CategoryIds, cancellationToken);
         if (categories.Count != request.CategoryIds.Distinct().Count())
         {
@@ -130,6 +136,11 @@ public class QuizService(
         var quiz = await quizRepository.GetByQuizIdAsync(id, cancellationToken);
         if (quiz == null)
             return;
+        
+        if (quiz.OwnerId != currentUser.UserId)
+        {
+            throw new ForbiddenException("delete this quiz");
+        }
 
         await quizRepository.DeleteAsync(id, cancellationToken);
     }
